Move client AMQP context to node-api so we can move the test serialization rule out.
