+++ serialization-tests/src/test/kotlin/net/corda/serialization/internal/amqp/DeserializeQueryableStateTest.kt
@@ -1,6 +1,6 @@
 package net.corda.serialization.internal.amqp
 
-import net.corda.client.rpc.internal.serialization.amqp.AMQPClientSerializationScheme
+import net.corda.nodeapi.internal.rpc.client.AMQPClientSerializationScheme
 import net.corda.core.identity.AbstractParty
 import net.corda.core.schemas.MappedSchema
 import net.corda.core.schemas.PersistentState
